+++ NEWS.md
@@ -60,6 +60,7 @@ For a more comprehensive changelog of the latest experimental code, see:
    - Added option to use Text To Speech for Sfinx.
 
  Cine:
+   - Added detection for Italian Amiga Operation Stealth.
    - Fixed vertically overflowing message boxes.
 
  Dreamweb: